Deal with the dataset type privilege for dataset operations
