+++ cdap-data-fabric/src/test/java/co/cask/cdap/data2/datafabric/dataset/RemoteDatasetFrameworkTest.java
@@ -30,9 +30,11 @@ import co.cask.cdap.common.metrics.NoOpMetricsCollectionService;
 import co.cask.cdap.data.dataset.SystemDatasetInstantiatorFactory;
 import co.cask.cdap.data.runtime.DynamicTransactionExecutorFactory;
 import co.cask.cdap.data2.datafabric.dataset.instance.DatasetInstanceManager;
+import co.cask.cdap.data2.datafabric.dataset.service.AuthorizationDatasetTypeService;
 import co.cask.cdap.data2.datafabric.dataset.service.DatasetInstanceService;
 import co.cask.cdap.data2.datafabric.dataset.service.DatasetService;
 import co.cask.cdap.data2.datafabric.dataset.service.DatasetTypeService;
+import co.cask.cdap.data2.datafabric.dataset.service.DefaultDatasetTypeService;
 import co.cask.cdap.data2.datafabric.dataset.service.executor.DatasetAdminOpHTTPHandler;
 import co.cask.cdap.data2.datafabric.dataset.service.executor.DatasetAdminService;
 import co.cask.cdap.data2.datafabric.dataset.service.executor.DatasetOpExecutor;
@@ -174,16 +176,19 @@ public class RemoteDatasetFrameworkTest extends AbstractDatasetFrameworkTest {
     DatasetInstanceManager instanceManager = new DatasetInstanceManager(txSystemClientService, txExecutorFactory,
                                                                         mdsFramework);
     PrivilegesManager privilegesManager = injector.getInstance(PrivilegesManager.class);
-    DatasetTypeService typeService = new DatasetTypeService(typeManager, namespaceQueryAdmin, namespacedLocationFactory,
-                                                            authorizationEnforcer, privilegesManager,
-                                                            authenticationContext, cConf, impersonator,
-                                                            txSystemClientService, mdsFramework, txExecutorFactory,
-                                                            DEFAULT_MODULES);
+    DatasetTypeService defaultService = new DefaultDatasetTypeService(typeManager, namespaceQueryAdmin,
+                                                                      namespacedLocationFactory, cConf, impersonator,
+                                                                      txSystemClientService, mdsFramework,
+                                                                      DEFAULT_MODULES);
+    DatasetTypeService typeService = new AuthorizationDatasetTypeService(defaultService, authorizationEnforcer,
+                                                                         privilegesManager, authenticationContext);
+
+
     DatasetOpExecutor opExecutor = new LocalDatasetOpExecutor(cConf, discoveryServiceClient, opExecutorService,
                                                               authenticationContext);
     DatasetInstanceService instanceService = new DatasetInstanceService(
-      typeService, instanceManager, opExecutor, exploreFacade, namespaceQueryAdmin, ownerAdmin, authorizationEnforcer,
-      privilegesManager, authenticationContext);
+      typeService, defaultService, instanceManager, opExecutor, exploreFacade, namespaceQueryAdmin, ownerAdmin,
+      authorizationEnforcer, privilegesManager, authenticationContext);
     instanceService.setAuditPublisher(inMemoryAuditPublisher);
 
     service = new DatasetService(cConf, discoveryService, discoveryServiceClient, metricsCollectionService,
